+++ jdm.properties
@@ -1,5 +1,5 @@
 #jDiskMark Properties File
-#Fri Sep 04 17:29:24 EDT 2020
+#Sun Nov 07 00:04:55 EDT 2021
 numOfFiles=50
 blockSequence=SEQUENTIAL
 readTest=true
@@ -7,7 +7,7 @@ blockSizeKb=128
 autoRemoveData=false
 autoReset=true
 showMaxMin=true
-locationDir=/Users/mcohen
+locationDir=C\:\\Users\\mcohen
 numOfBlocks=256
 writeSyncEnable=true
 writeTest=true
