--- src/META-INF/persistence.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<persistence xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" version="2.1"
-             xmlns="http://xmlns.jcp.org/xml/ns/persistence"
-             xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/persistence http://xmlns.jcp.org/xml/ns/persistence/persistence_2_1.xsd">
-    <persistence-unit name="jDiskMarkPU" transaction-type="RESOURCE_LOCAL">
-        <provider>org.eclipse.persistence.jpa.PersistenceProvider</provider>
-        <exclude-unlisted-classes>false</exclude-unlisted-classes>
-        <properties>
-            <property name="eclipselink.logging.level" value="FINE"/>
-            <property name="jakarta.persistence.jdbc.url" value="jdbc:derby:derbyDB;create=true"/>
-            <property name="jakarta.persistence.schema-generation.database.action" value="create"/>
-            <property name="eclipselink.ddl-generation" value="create-or-extend-tables"/>
-        </properties>
-    </persistence-unit>
-</persistence>
