--- src/edu/touro/mco152/bm/persist/package-info.java
@@ -1,1 +1,0 @@
-package edu.touro.mco152.bm.persist;
